feat/Refactoring
